feat: support sorted index build options
