+++ crates/paimon/src/btree/mod.rs
@@ -47,6 +47,7 @@ pub(crate) mod var_len;
 mod writer;
 
 pub use block::BlockCompressionType;
+pub(crate) use block::{compress_block, compute_crc32, decompress_block};
 pub use footer::BTreeFileFooter;
 pub use key_serde::{make_key_comparator, serialize_datum};
 pub use meta::BTreeIndexMeta;
