+++ crates/paimon/src/btree/tests.rs
@@ -384,23 +384,34 @@ fn test_meta_serialization() {
 }
 
 #[tokio::test]
-async fn test_zstd_compression() {
-    let buf = VecFileWrite::new();
-    let mut writer = BTreeIndexWriter::new(Box::new(buf.clone()), 64, BlockCompressionType::Zstd);
+async fn test_supported_compressions() {
+    for compression_type in [
+        BlockCompressionType::Zstd,
+        BlockCompressionType::Lz4,
+        BlockCompressionType::Lzo,
+    ] {
+        let buf = VecFileWrite::new();
+        let mut writer = BTreeIndexWriter::with_compression_level(
+            Box::new(buf.clone()),
+            64,
+            compression_type,
+            7,
+        );
 
-    for i in 0..100 {
-        writer.write(Some(&int_key(i)), i as i64).await.unwrap();
+        for i in 0..100 {
+            writer.write(Some(&int_key(i)), i as i64).await.unwrap();
+        }
+
+        let result = writer.finish().await.unwrap();
+        let reader = write_and_open(&buf, &result, int_cmp).await;
+
+        let all = reader.all_non_null_rows().await.unwrap();
+        assert_eq!(all.len(), 100, "{compression_type:?}");
+
+        let bm = reader.query_equal(&int_key(50)).await.unwrap();
+        assert_eq!(bm.len(), 1, "{compression_type:?}");
+        assert!(bm.contains(50), "{compression_type:?}");
     }
-
-    let result = writer.finish().await.unwrap();
-    let reader = write_and_open(&buf, &result, int_cmp).await;
-
-    let all = reader.all_non_null_rows().await.unwrap();
-    assert_eq!(all.len(), 100);
-
-    let bm = reader.query_equal(&int_key(50)).await.unwrap();
-    assert_eq!(bm.len(), 1);
-    assert!(bm.contains(50));
 }
 
 #[tokio::test]
@@ -664,25 +675,14 @@ async fn test_java_compat_varchar_no_compress() {
 }
 
 #[tokio::test]
-async fn test_java_compat_int_lz4_unsupported() {
-    let data = Bytes::from(load_testdata("btree_int_100_lz4.bin"));
-    let file_size = data.len() as u64;
+async fn test_java_compat_int_lz4() {
     let meta = BTreeIndexMeta::new(Some(le_int_key(0)), Some(le_int_key(198)), false);
-    let reader_result =
-        BTreeIndexReader::open(Box::new(BytesFileRead(data)), file_size, &meta, le_int_cmp).await;
+    let reader = open_testdata("btree_int_100_lz4.bin", &meta, le_int_cmp).await;
 
-    match reader_result {
-        Err(e) => {
-            assert!(
-                e.to_string().contains("not supported") || e.to_string().contains("Unsupported"),
-                "Expected unsupported compression error, got: {e}"
-            );
-        }
-        Ok(reader) => {
-            // If reader creation succeeded (index block not compressed),
-            // data block read should fail
-            let result = reader.query_equal(&le_int_key(0)).await;
-            assert!(result.is_err(), "LZ4 data block read should fail");
-        }
-    }
+    let all = reader.all_non_null_rows().await.unwrap();
+    assert_eq!(all.len(), 100);
+
+    let bm = reader.query_equal(&le_int_key(100)).await.unwrap();
+    assert_eq!(bm.len(), 1);
+    assert!(bm.contains(50));
 }
